fix(Group): handle visibility for nested groups
When a group has another group as only child, its slot may not have an associated
`componentInstance`. In this case we need to look in its children for componentInstances to check for visibility.

This was found in async_paraview when using the stream tracer filter with the following layout.

```
<SourceProxy StreamTracer ...>
  <PropertyGroup label="Seeds">
         <Property name="Source"/>
         <PropertyGroup label="Line Parameters" panel_widget="InteractiveLine">
            ...
        </PropertyGroup>
  </PropertyGroup>
</SourceProxy>
```

Since we do not have `InteractiveLine` widget yet and we replace it with a group and thus a nested group is created.