fix(commit): ensure committed proxies have pending properties.
The very first time we commit a change is committed via a simput widget
`_pending_changeset` holds the values of the edited widget while the `_ids`
holds all the proxies that haven't been committed so far. This yields an error
if there is a proxy in `_ids` but not in `_pending_changeset`.

The previous approach would assume that any non-committed proxy had pending
changes which is not true (at least in Async ParaView).  In Async ParaView,
proxies created without the UI like a render view has all its properties
already setup by the time the SimputWidget for the view is created so even
though is a "new" proxy, from simput's point of view there are no
pending_changes.

Other applications consuming `trame-simput` should not be affected negatively
since we just protect existing code that accessing a dictionary for missing
keys without any other restriction.